Refactor uvicorn into dev module
